+++ modules/apps/bookmarks/bookmarks-web/src/main/resources/META-INF/resources/init.jsp
@@ -44,7 +44,7 @@ page import="com.liferay.bookmarks.service.BookmarksFolderServiceUtil" %><%@
 page import="com.liferay.bookmarks.service.permission.BookmarksEntryPermissionChecker" %><%@
 page import="com.liferay.bookmarks.service.permission.BookmarksFolderPermissionChecker" %><%@
 page import="com.liferay.bookmarks.web.dao.search.BookmarksResultRowSplitter" %><%@
-page import="com.liferay.bookmarks.web.portlet.toolbar.item.BookmarksPortletToolbarContributor" %><%@
+page import="com.liferay.bookmarks.web.portlet.toolbar.contributor.BookmarksPortletToolbarContributor" %><%@
 page import="com.liferay.bookmarks.web.portlet.util.BookmarksUtil" %><%@
 page import="com.liferay.bookmarks.web.search.EntriesChecker" %><%@
 page import="com.liferay.frontend.taglib.servlet.taglib.AddMenuItem" %><%@
