+++ modules/apps/message-boards/message-boards-web/src/main/java/com/liferay/message/boards/web/portlet/toolbar/contributor/MBPortletToolbarContributor.java
@@ -12,28 +12,17 @@
  * details.
  */
 
-package com.liferay.bookmarks.web.portlet.toolbar.item;
+package com.liferay.message.boards.web.portlet.toolbar.contributor;
 
-import com.liferay.bookmarks.configuration.BookmarksGroupServiceOverriddenConfiguration;
-import com.liferay.bookmarks.constants.BookmarksConstants;
-import com.liferay.bookmarks.constants.BookmarksPortletKeys;
-import com.liferay.bookmarks.constants.BookmarksWebKeys;
-import com.liferay.bookmarks.exception.NoSuchFolderException;
-import com.liferay.bookmarks.model.BookmarksFolder;
-import com.liferay.bookmarks.model.BookmarksFolderConstants;
-import com.liferay.bookmarks.service.BookmarksFolderService;
-import com.liferay.portal.kernel.bean.BeanParamUtil;
+import com.liferay.message.boards.web.constants.MBPortletKeys;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.module.configuration.ConfigurationException;
-import com.liferay.portal.kernel.module.configuration.ConfigurationFactoryUtil;
 import com.liferay.portal.kernel.portlet.toolbar.contributor.PortletToolbarContributor;
 import com.liferay.portal.kernel.servlet.taglib.ui.Menu;
 import com.liferay.portal.kernel.servlet.taglib.ui.MenuItem;
 import com.liferay.portal.kernel.servlet.taglib.ui.URLMenuItem;
-import com.liferay.portal.kernel.settings.GroupServiceSettingsLocator;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.security.permission.ActionKeys;
 import com.liferay.portal.security.permission.BaseModelPermissionChecker;
@@ -42,7 +31,9 @@ import com.liferay.portal.theme.PortletDisplay;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PortalUtil;
 import com.liferay.portlet.PortletURLFactoryUtil;
-import com.liferay.portlet.documentlibrary.model.DLFolderConstants;
+import com.liferay.portlet.messageboards.model.MBCategory;
+import com.liferay.portlet.messageboards.model.MBCategoryConstants;
+import com.liferay.portlet.messageboards.util.MBUtil;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -59,16 +50,13 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	immediate = true,
 	property = {
-		"javax.portlet.name=" + BookmarksPortletKeys.BOOKMARKS,
-		"mvc.render.command.name=-", "mvc.render.command.name=/bookmarks/view"
+		"javax.portlet.name=" + MBPortletKeys.MESSAGE_BOARDS,
+		"mvc.render.command.name=-",
+		"mvc.render.command.name=/message_boards/view"
 	},
-	service = {
-		BookmarksPortletToolbarContributor.class,
-		PortletToolbarContributor.class
-	}
+	service = {MBPortletToolbarContributor.class, PortletToolbarContributor.class}
 )
-public class BookmarksPortletToolbarContributor
-	implements PortletToolbarContributor {
+public class MBPortletToolbarContributor implements PortletToolbarContributor {
 
 	@Override
 	public List<Menu> getPortletTitleMenus(PortletRequest portletRequest) {
@@ -87,26 +75,30 @@ public class BookmarksPortletToolbarContributor
 		return menus;
 	}
 
-	protected void addPortletTitleAddBookmarkMenuItem(
-			List<MenuItem> menuItems, BookmarksFolder folder,
+	protected void addPortletTitleAddCategoryMenuItem(
+			List<MenuItem> menuItems, long categoryId,
 			ThemeDisplay themeDisplay, PortletRequest portletRequest)
 		throws PortalException {
 
-		long folderId = _getFolderId(folder);
-
 		if (!containsPermission(
 				themeDisplay.getPermissionChecker(),
-				themeDisplay.getScopeGroupId(), folderId,
-				ActionKeys.ADD_ENTRY)) {
+				themeDisplay.getScopeGroupId(), categoryId,
+				ActionKeys.ADD_CATEGORY)) {
 
 			return;
 		}
 
 		URLMenuItem urlMenuItem = new URLMenuItem();
 
+		String key = "category[message-board]";
+
+		if (categoryId != MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID) {
+			key = "subcategory[message-board]";
+		}
+
 		urlMenuItem.setLabel(
 			LanguageUtil.get(
-				PortalUtil.getHttpServletRequest(portletRequest), "bookmark"));
+				PortalUtil.getHttpServletRequest(portletRequest), key));
 
 		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
 
@@ -115,27 +107,25 @@ public class BookmarksPortletToolbarContributor
 			PortletRequest.RENDER_PHASE);
 
 		portletURL.setParameter(
-			"mvcRenderCommandName", "/bookmarks/edit_entry");
+			"mvcRenderCommandName", "/message_boards/edit_category");
 		portletURL.setParameter(
 			"redirect", PortalUtil.getCurrentURL(portletRequest));
-		portletURL.setParameter("folderId", String.valueOf(folderId));
+		portletURL.setParameter("parentCategoryId", String.valueOf(categoryId));
 
 		urlMenuItem.setURL(portletURL.toString());
 
 		menuItems.add(urlMenuItem);
 	}
 
-	protected void addPortletTitleAddFolderMenuItem(
-			List<MenuItem> menuItems, BookmarksFolder folder,
+	protected void addPortletTitleAddMessageMenuItem(
+			List<MenuItem> menuItems, long categoryId,
 			ThemeDisplay themeDisplay, PortletRequest portletRequest)
 		throws PortalException {
 
-		long folderId = _getFolderId(folder);
-
 		if (!containsPermission(
 				themeDisplay.getPermissionChecker(),
-				themeDisplay.getScopeGroupId(), folderId,
-				ActionKeys.ADD_FOLDER)) {
+				themeDisplay.getScopeGroupId(), categoryId,
+				ActionKeys.ADD_MESSAGE)) {
 
 			return;
 		}
@@ -144,8 +134,7 @@ public class BookmarksPortletToolbarContributor
 
 		urlMenuItem.setLabel(
 			LanguageUtil.get(
-				PortalUtil.getHttpServletRequest(portletRequest),
-				(folder != null) ? "subfolder" : "folder"));
+				PortalUtil.getHttpServletRequest(portletRequest), "thread"));
 
 		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
 
@@ -154,10 +143,10 @@ public class BookmarksPortletToolbarContributor
 			PortletRequest.RENDER_PHASE);
 
 		portletURL.setParameter(
-			"mvcRenderCommandName", "/bookmarks/edit_folder");
+			"mvcRenderCommandName", "/message_boards/edit_message");
 		portletURL.setParameter(
 			"redirect", PortalUtil.getCurrentURL(portletRequest));
-		portletURL.setParameter("parentFolderId", String.valueOf(folderId));
+		portletURL.setParameter("mbCategoryId", String.valueOf(categoryId));
 
 		urlMenuItem.setURL(portletURL.toString());
 
@@ -185,31 +174,31 @@ public class BookmarksPortletToolbarContributor
 		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
 
-		BookmarksFolder folder = _getFolder(themeDisplay, portletRequest);
+		long categoryId = _getCategoryId(portletRequest);
 
 		List<MenuItem> menuItems = new ArrayList<>();
 
 		try {
-			addPortletTitleAddFolderMenuItem(
-				menuItems, folder, themeDisplay, portletRequest);
+			addPortletTitleAddCategoryMenuItem(
+				menuItems, categoryId, themeDisplay, portletRequest);
 		}
 		catch (PortalException pe) {
 			_log.error("Unable to add folder menu item", pe);
 		}
 
 		try {
-			addPortletTitleAddBookmarkMenuItem(
-				menuItems, folder, themeDisplay, portletRequest);
+			addPortletTitleAddMessageMenuItem(
+				menuItems, categoryId, themeDisplay, portletRequest);
 		}
 		catch (PortalException pe) {
-			_log.error("Unable to add bookmark menu item", pe);
+			_log.error("Unable to add shortcut menu item", pe);
 		}
 
 		return menuItems;
 	}
 
 	@Reference(
-		target = "(model.class.name=com.liferay.bookmarks.model.BookmarksFolder)",
+		target = "(model.class.name=com.liferay.portlet.messageboards.model.MBCategory)",
 		unbind = "-"
 	)
 	protected void setBaseModelPermissionChecker(
@@ -218,73 +207,17 @@ public class BookmarksPortletToolbarContributor
 		_baseModelPermissionChecker = baseModelPermissionChecker;
 	}
 
-	@Reference(unbind = "-")
-	protected void setBookmarksFolderService(
-		BookmarksFolderService bookmarksFolderService) {
+	private long _getCategoryId(PortletRequest portletRequest) {
+		MBCategory category = (MBCategory)portletRequest.getAttribute(
+			WebKeys.MESSAGE_BOARDS_CATEGORY);
 
-		_bookmarksFolderService = bookmarksFolderService;
-	}
-
-	private BookmarksFolder _getFolder(
-		ThemeDisplay themeDisplay, PortletRequest portletRequest) {
-
-		long rootFolderId = BookmarksFolderConstants.DEFAULT_PARENT_FOLDER_ID;
-
-		try {
-			BookmarksGroupServiceOverriddenConfiguration
-				bookmarksGroupServiceOverriddenConfiguration =
-					ConfigurationFactoryUtil.getConfiguration(
-						BookmarksGroupServiceOverriddenConfiguration.class,
-						new GroupServiceSettingsLocator(
-							themeDisplay.getScopeGroupId(),
-							BookmarksConstants.SERVICE_NAME));
-
-			rootFolderId =
-				bookmarksGroupServiceOverriddenConfiguration.rootFolderId();
-		}
-		catch (ConfigurationException ce) {
-			_log.error(
-				"Unable to obtain bookmarks root folder ID for group " +
-					themeDisplay.getScopeGroupId());
-		}
-
-		BookmarksFolder folder = (BookmarksFolder)portletRequest.getAttribute(
-			BookmarksWebKeys.BOOKMARKS_FOLDER);
-
-		long folderId = BeanParamUtil.getLong(
-			folder, portletRequest, "folderId", rootFolderId);
-
-		if ((folder == null) &&
-			(folderId != BookmarksFolderConstants.DEFAULT_PARENT_FOLDER_ID)) {
-
-			try {
-				folder = _bookmarksFolderService.getFolder(folderId);
-			}
-			catch (NoSuchFolderException nsfe) {
-				folder = null;
-			}
-			catch (PortalException pe) {
-				_log.error(pe, pe);
-			}
-		}
-
-		return folder;
-	}
-
-	private long _getFolderId(BookmarksFolder folder) {
-		long folderId = DLFolderConstants.DEFAULT_PARENT_FOLDER_ID;
-
-		if (folder != null) {
-			folderId = folder.getFolderId();
-		}
-
-		return folderId;
+		return MBUtil.getCategoryId(
+			PortalUtil.getHttpServletRequest(portletRequest), category);
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		BookmarksPortletToolbarContributor.class);
+		MBPortletToolbarContributor.class);
 
 	private volatile BaseModelPermissionChecker _baseModelPermissionChecker;
-	private volatile BookmarksFolderService _bookmarksFolderService;
 
 }
